Move RepresentativenessData typings into hook
