+++ front/app/modules/commercial/representativeness/admin/hooks/useReferenceData.ts
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 
 // services
 import {
@@ -13,14 +13,34 @@ import {
   TCustomFieldCode,
 } from 'modules/commercial/user_custom_fields/services/userCustomFields';
 
+export interface RepresentativenessRow {
+  name: string;
+  actualPercentage: number;
+  referencePercentage: number;
+  actualNumber: number;
+  referenceNumber: number;
+}
+
+export type RepresentativenessData = RepresentativenessRow[];
+
 const getStream = (code: TCustomFieldCode | null) => {
   if (code === null) return usersByRegFieldStream;
   if (code === 'gender') return usersByGenderStream;
   if (code === 'domicile') return usersByDomicileStream;
+  return undefined;
 };
 
-function useReferenceData(field: IUserCustomFieldData) {
-  const [referenceData, useReferenceData] = useState();
+function useReferenceData(field: IUserCustomFieldData, projectId?: string) {
+  const [referenceData, setReferenceData] = useState();
+  const code = field.attributes.code;
+
+  useEffect(() => {
+    const stream = getStream(code);
+    if (!stream) {
+      setReferenceData(null);
+      return;
+    }
+  }, [code]);
 }
 
 export default useReferenceData;
